Derivation node counts converge for GLLHashPool and GLLModal

+++ src/uk/ac/rhul/cs/csle/art/util/lexicalisations/AbstractLexicalisations.java
@@ -22,7 +22,7 @@ public abstract class AbstractLexicalisations implements DisplayInterface {
 
   public abstract boolean valid();
 
-  public abstract int cardinality();
+  public abstract long cardinality();
 
   public abstract boolean isDeterministic();
